backend_and_ML/requirements.txt
fastapi
uvicorn
python-dotenv
motor
bcrypt
httpx
pydantic